back-end-v2/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation et sérialisation
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# LLM Providers
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1

# Validation JSON
jsonschema==4.21.1

# Tâches asynchrones
celery==5.3.6
redis==5.0.1

# Stockage S3
boto3==1.34.34

# Métriques et monitoring
prometheus-client==0.19.0

# HTTP client
httpx==0.26.0

# Configuration
python-dotenv==1.0.0

# Logging structuré
structlog==24.1.0

# Tests
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

